// frontend/src/gen/api/src/models/SearchBarUserDetailDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SearchBarUserDetailDto
 */
export interface SearchBarUserDetailDto {
    /**
     * 
     * @type {string}
     * @memberof SearchBarUserDetailDto
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchBarUserDetailDto
     */
    nickName?: string;
    /**
     * 
     * @type {string}
     * @memberof SearchBarUserDetailDto
     */
    avatarUrl?: string;
}

/**
 * Check if a given object implements the SearchBarUserDetailDto interface.
 */
export function instanceOfSearchBarUserDetailDto(value: object): boolean {
    return true;
}

export function SearchBarUserDetailDtoFromJSON(json: any): SearchBarUserDetailDto {
    return SearchBarUserDetailDtoFromJSONTyped(json, false);
}

export function SearchBarUserDetailDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchBarUserDetailDto {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'] == null ? undefined : json['userId'],
        'nickName': json['nickName'] == null ? undefined : json['nickName'],
        'avatarUrl': json['avatarUrl'] == null ? undefined : json['avatarUrl'],
    };
}

export function SearchBarUserDetailDtoToJSON(value?: SearchBarUserDetailDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'userId': value['userId'],
        'nickName': value['nickName'],
        'avatarUrl': value['avatarUrl'],
    };
}